+++ changelog.md
@@ -25,6 +25,7 @@ Codion Change Log
 - EntityComboBoxModel.ForeignKeyComboBoxModelFactory and ForeignKeyComboBoxModelLinker removed, ForeignKeyFilter.link() added.
 - EntityComboBoxModel.Builder added, related refactoring.
 - EntityComboBoxModel.ForeignKeyFilter.builder() added.
+- EntityComboBoxModel.ForeignKeyFilter.builder() now sets includeNull() to true if the foreign key is nullable.
 ### is.codion.swing.framework.ui
 - EntityTablePanel.Config.cellRenderer() and cellEditor() added.
 - EntityComboBox.model() added for consistency.